requirements.txt
matplotlib==3.5.3
numpy==1.21.6
Pillow==9.4.0
progress==1.6
scipy==1.10.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
